Trace back is seen on populate.log
ConfigFileValueError: invalid literal for
int() with base 10: 'rabbit_port'. At the very beginning of Valet
deployment, /usr/etc/valet/valet.conf file is getting read instead
of /etc/valet/valet.conf. Here the value of port is expected to be
an int and not string. Changing the value to int.

Change-Id: I820c5f44a76d5a4a313d896b50242071b37eb6b2