Cap ostestr at below 1.0.0
The latest update to ostestr which introduced using stestr
instead of testr introduced something which is currently causing
the py27 unit tests to fail with ArgsAlreadyParsedError. This
change caps ostestr at <1.0.0 to allow the unit tests to run
successfully while we figure out what is causing the issue
when ran with stestr.

Change-Id: I96bc0e443a630f28ca06185a88e2a5d4e8e4c5a8